24 modul zarzadzania rynkiem (#42)
* Dodano dla GM'a decyzje do zatwierdzenia oraz zdarzenia losowe

* Zmiany w bazie danych: Dodano tabelę GameEvent i przykładowe 3 eventy na każdy możliwy rodzaj eventu.

* Dodanie dłuższego opisu kart po ich wybraniu i dodanie switcha przełączającego decyzje do zatwierdzenia i historie decyzji

* Dodano działający panel historii po stronie admina

* Dodanie działające sugerowanie kart przez gracza i zatwierdzanie/odrzucanie ich przez admina

* Dodano przełącznik miedzy kartami i przedmiotami oraz dodano odswiezanie historii i decyzji do zatwierdzenia po zagraniu karty przez gracza

* Dodano działające zagrywanie kart decyzji/przedmiotów po stronie admina

* Dodanie działających eventów, wyświetlanie informacji o nich w historii decyzji i oznaczanie kart, które są zagrane w evencie

* Zmiana nazewnicta i sciezek

* Dodaną nową metodę SignalR do odświeżania historii decyzji i zatwierdzeń na panelu admina (Na spokojnie można używać tego do innych metod)

PS Trzeba zrobić "npm install @microsoft/signalr" na froncie by to działało, przez gitignore nie mogłem dać tego do commita

* Dodanie SignalR do odswiezania historii decyzji w panelu gracza

* Dziala dynamiczne odświeżanie planszy gracza

* Dynamiczne odswiezanie planszy w widoku konkretnego stolu

* Dodano panel zarzadzania stolem

* Poprawiono wagi procesów do obliczania poycji pionka drużyny

* Plansza w module zarządzania rynkiem wyświetla się prawidłowo, pozycje pionków wyświetlają się prawidłowo, poprawiono ładowanie opisów wydarzeń na panelu sterowania stołem oraz w module

---------

+++ backend/Data/DecisionWeight.cs
@@ -11,7 +11,7 @@ namespace backend.Data
         public Deck Deck { get; set; } = null!;
         public int WeightX { get; set; }
         public int WeightY { get; set; }
-        public int BoosterX { get; set; }
-        public int BoosterY { get; set; }
+        public double BoosterX { get; set; }
+        public double BoosterY { get; set; }
     }
 }
